data

+++ app/src/main/java/Models/Guitars.kt
@@ -3,9 +3,9 @@ package Models
 data class Guitars(
     val name: String,
     val imgSrc: Int,
-    val price: Float,
+    val price: Int,
     val description: String,
-    val scaleLength: Float,
+    val scaleLength: Double,
     val body: String,
-    val rating: Float
+    val rating: Double
 )
